Reconcile PicoMUTX into the canonical system
